refactor: Order 패키지에서 OrderTable 분리
