+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -9,13 +9,13 @@ import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.order.domain.OrderedMenu;
 import kitchenpos.order.domain.Price;
 import kitchenpos.order.exception.NotFoundException;
 import kitchenpos.order.repository.OrderLineItemRepository;
 import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.order.repository.OrderTableRepository;
+import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.order.ui.dto.request.OrderCreateRequest;
 import kitchenpos.order.ui.dto.request.OrderLineItemDto;
 import kitchenpos.order.ui.dto.request.OrderStatusChangeRequest;
