--- src/main/java/kitchenpos/order/domain/TableGroup.java
@@ -1,82 +1,0 @@
-package kitchenpos.order.domain;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
-
-@Entity
-public class TableGroup {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private LocalDateTime createdDate;
-
-    @OneToMany(fetch = FetchType.LAZY)
-    @JoinColumn(name = "table_group_id")
-    private List<OrderTable> orderTables;
-
-    public TableGroup() {
-    }
-
-    public TableGroup(final LocalDateTime createdDate, final List<OrderTable> orderTables) {
-        this.createdDate = createdDate;
-        this.orderTables = orderTables;
-    }
-
-    private TableGroup(final Builder builder) {
-        this.id = builder.id;
-        this.createdDate = builder.createdDate;
-        this.orderTables = builder.orderTables;
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public List<OrderTable> getOrderTables() {
-        return orderTables;
-    }
-
-    public static class Builder {
-        private Long id;
-        private LocalDateTime createdDate;
-        private List<OrderTable> orderTables;
-
-        public Builder id(final Long id) {
-            this.id = id;
-            return this;
-        }
-
-        public Builder createdDate(final LocalDateTime createdDate) {
-            this.createdDate = createdDate;
-            return this;
-        }
-
-        public Builder orderTables(final List<OrderTable> orderTables) {
-            this.orderTables = orderTables;
-            return this;
-        }
-
-        public TableGroup build() {
-            return new TableGroup(this);
-        }
-    }
-}
