--- src/main/java/kitchenpos/order/repository/OrderTableRepository.java
@@ -1,12 +1,0 @@
-package kitchenpos.order.repository;
-
-import java.util.List;
-import kitchenpos.order.domain.OrderTable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
-
-    List<OrderTable> findAllByIdIn(List<Long> ids);
-
-    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
-}
