--- src/main/java/kitchenpos/order/repository/TableGroupRepository.java
@@ -1,7 +1,0 @@
-package kitchenpos.order.repository;
-
-import kitchenpos.order.domain.TableGroup;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
-}
