--- src/main/java/kitchenpos/order/ui/TableGroupRestController.java
@@ -1,38 +1,0 @@
-package kitchenpos.order.ui;
-
-import java.net.URI;
-import javax.validation.Valid;
-import kitchenpos.order.application.TableGroupService;
-import kitchenpos.order.ui.dto.request.TableGroupCreateRequest;
-import kitchenpos.order.ui.dto.response.TableGroupCreateResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TableGroupRestController {
-
-    private final TableGroupService tableGroupService;
-
-    public TableGroupRestController(final TableGroupService tableGroupService) {
-        this.tableGroupService = tableGroupService;
-    }
-
-    @PostMapping("/api/table-groups")
-    public ResponseEntity<TableGroupCreateResponse> create(@RequestBody @Valid final TableGroupCreateRequest request) {
-        final TableGroupCreateResponse response = tableGroupService.create(request);
-        final URI uri = URI.create("/api/table-groups/" + response.getId());
-        return ResponseEntity.created(uri)
-                .body(response);
-    }
-
-    @DeleteMapping("/api/table-groups/{tableGroupId}")
-    public ResponseEntity<Void> ungroup(@PathVariable final Long tableGroupId) {
-        tableGroupService.ungroup(tableGroupId);
-        return ResponseEntity.noContent()
-                .build();
-    }
-}
