--- src/main/java/kitchenpos/order/ui/dto/request/OrderTableChangeEmptyRequest.java
@@ -1,20 +1,0 @@
-package kitchenpos.order.ui.dto.request;
-
-import javax.validation.constraints.NotNull;
-
-public class OrderTableChangeEmptyRequest {
-
-    @NotNull
-    private Boolean empty;
-
-    public OrderTableChangeEmptyRequest() {
-    }
-
-    public OrderTableChangeEmptyRequest(final Boolean empty) {
-        this.empty = empty;
-    }
-
-    public Boolean getEmpty() {
-        return empty;
-    }
-}
