--- src/main/java/kitchenpos/order/ui/dto/request/OrderTableChangeNumberOfGuestsRequest.java
@@ -1,20 +1,0 @@
-package kitchenpos.order.ui.dto.request;
-
-import javax.validation.constraints.Min;
-
-public class OrderTableChangeNumberOfGuestsRequest {
-
-    @Min(0)
-    private int numberOfGuests;
-
-    public OrderTableChangeNumberOfGuestsRequest() {
-    }
-
-    public OrderTableChangeNumberOfGuestsRequest(final int numberOfGuests) {
-        this.numberOfGuests = numberOfGuests;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-}
