--- src/main/java/kitchenpos/order/ui/dto/response/OrderTableChangeEmptyResponse.java
@@ -1,43 +1,0 @@
-package kitchenpos.order.ui.dto.response;
-
-import kitchenpos.order.domain.OrderTable;
-
-public class OrderTableChangeEmptyResponse {
-
-    private Long id;
-    private Long tableGroupId;
-    private int numberOfGuests;
-    private boolean empty;
-
-    public OrderTableChangeEmptyResponse() {
-    }
-
-    public OrderTableChangeEmptyResponse(final Long id, final Long tableGroupId, final int numberOfGuests,
-                                         final boolean empty) {
-        this.id = id;
-        this.tableGroupId = tableGroupId;
-        this.numberOfGuests = numberOfGuests;
-        this.empty = empty;
-    }
-
-    public static OrderTableChangeEmptyResponse from(final OrderTable orderTable) {
-        return new OrderTableChangeEmptyResponse(orderTable.getId(), orderTable.getTableGroupId(),
-                orderTable.getNumberOfGuests(), orderTable.isEmpty());
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Long getTableGroupId() {
-        return tableGroupId;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-}
