+++ src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -1,0 +1,122 @@
+package kitchenpos.table.domain;
+
+import java.util.Objects;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class OrderTable {
+
+    private static final String ALREADY_GROUP_ERROR_MESSAGE = "이미 단체지정되어 있는 테이블은 단체지정 할 수 없습니다";
+    private static final String CHANGE_EMPTY_ERROR_MESSAGE = "단체 지정되어있는 테이블은 테이블이 비어있는지 여부를 수정할 수 없습니다.";
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "table_group_id")
+    private Long tableGroupId;
+
+    @Column(nullable = false)
+    private int numberOfGuests;
+
+    @Column(nullable = false)
+    private boolean empty;
+
+    public OrderTable() {
+    }
+
+    public OrderTable(Builder builder) {
+        this.id = builder.id;
+        this.tableGroupId = builder.tableGroupId;
+        this.numberOfGuests = builder.numberOfGuests;
+        this.empty = builder.empty;
+    }
+
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    public void changeEmpty(final boolean empty) {
+        if (Objects.nonNull(tableGroupId)) {
+            throw new IllegalArgumentException(OrderTable.CHANGE_EMPTY_ERROR_MESSAGE);
+        }
+        this.empty = empty;
+    }
+
+    public void changeNumberOfGuests(final int numberOfGuests) {
+        if (empty) {
+            throw new IllegalArgumentException();
+        }
+        this.numberOfGuests = numberOfGuests;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Long getTableGroupId() {
+        return tableGroupId;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+
+    public void group(final Long tableGroupId) {
+        this.tableGroupId = tableGroupId;
+        this.empty = false;
+    }
+
+    public void ungroup() {
+        this.tableGroupId = null;
+        this.empty = false;
+    }
+
+    public void validateCanGroup() {
+        if (Objects.nonNull(tableGroupId)) {
+            throw new IllegalArgumentException(ALREADY_GROUP_ERROR_MESSAGE);
+        }
+        if (!empty) {
+            throw new IllegalArgumentException(CHANGE_EMPTY_ERROR_MESSAGE);
+        }
+    }
+
+    public static class Builder {
+        private Long id;
+        private Long tableGroupId;
+        private int numberOfGuests;
+        private boolean empty;
+
+        public Builder id(final Long id) {
+            this.id = id;
+            return this;
+        }
+
+        public Builder tableGroupId(final Long tableGroupId) {
+            this.tableGroupId = tableGroupId;
+            return this;
+        }
+
+        public Builder numberOfGuests(final int numberOfGuests) {
+            this.numberOfGuests = numberOfGuests;
+            return this;
+        }
+
+        public Builder empty(final boolean empty) {
+            this.empty = empty;
+            return this;
+        }
+
+        public OrderTable build() {
+            return new OrderTable(this);
+        }
+    }
+}
