+++ src/main/java/kitchenpos/table/repository/OrderTableRepository.java
@@ -1,0 +1,12 @@
+package kitchenpos.table.repository;
+
+import java.util.List;
+import kitchenpos.table.domain.OrderTable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
+
+    List<OrderTable> findAllByIdIn(List<Long> ids);
+
+    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
+}
