+++ src/main/java/kitchenpos/table/repository/TableGroupRepository.java
@@ -1,0 +1,7 @@
+package kitchenpos.table.repository;
+
+import kitchenpos.table.domain.TableGroup;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
+}
