+++ src/main/java/kitchenpos/table/ui/dto/request/OrderTableChangeEmptyRequest.java
@@ -1,0 +1,20 @@
+package kitchenpos.table.ui.dto.request;
+
+import javax.validation.constraints.NotNull;
+
+public class OrderTableChangeEmptyRequest {
+
+    @NotNull
+    private Boolean empty;
+
+    public OrderTableChangeEmptyRequest() {
+    }
+
+    public OrderTableChangeEmptyRequest(final Boolean empty) {
+        this.empty = empty;
+    }
+
+    public Boolean getEmpty() {
+        return empty;
+    }
+}
