+++ src/main/java/kitchenpos/table/ui/dto/response/TableCreateResponse.java
@@ -1,0 +1,42 @@
+package kitchenpos.table.ui.dto.response;
+
+import kitchenpos.table.domain.OrderTable;
+
+public class TableCreateResponse {
+
+    private Long id;
+    private Long tableGroupId;
+    private int numberOfGuests;
+    private boolean empty;
+
+    public TableCreateResponse() {
+    }
+
+    private TableCreateResponse(final Long id, final Long tableGroupId, final int numberOfGuests, final boolean empty) {
+        this.id = id;
+        this.tableGroupId = tableGroupId;
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public static TableCreateResponse from(final OrderTable orderTable) {
+        return new TableCreateResponse(orderTable.getId(), orderTable.getTableGroupId(), orderTable.getNumberOfGuests(),
+                orderTable.isEmpty());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Long getTableGroupId() {
+        return tableGroupId;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+}
